Adds `BeAssignableTo` implementation for `Type` objects.
In the absence of this new hiding definition, FA would try to assert the Type, `Type`, was assignable to the expectd type; an assertion having no real value.

This also updates the formatting on the specs of the hidden method.

This also corrects a slight inaccuracy in the failure message of the hidden method.