Adds new extension methods to make existing filtering of `Type`s functionality more accessible.
This is done on the premise that these methods are rather useful; but, hard to find.

These extension methods are still on a fairly specialized type so they  should not cause much of a pollution problem.  Even then these might be fairly common things to want to do on that type so I still think it would not be pollution.