Assuming types that override Equals to have value semantics (#740)
Replaced all mechanisms and options to control whether or not an object has value semantics by a single algorithm. Any object that overrides object.Equals and is not an anonymous type or tuple is treated as a value type, unless ComparingByMembers<T>() is used. Also removes the global AssertionOptions.IsValueType