Switch to light AAA syntax
Until now, or at least for a long time, we have used a quite heavy comment
style to separate the three phases of a unit test following the AAA pattern.
While it provided a *very* clear separation, it also added a high
comment-to-code ratio. Switching to the more commonly used single line
comment style reduces the test files by ~23% number of lines.