Use a counting graph to keep track of xpath index
Currently we use a stack to keep track of the location and a dictionary to keep track of the xpath index.
That has multiple downsides:

* We need to keep two separate data structure in sync.
* We're a lot of string allocations on each iterations when computing path as key for the dictionary.
* When creating the failure message we're doing lookups in two data structures

Instead we can combine their purposes in a single graph where each node has a `Name` and a `Count`, where the latter is updated each time we visit a node with the same path.

On each xml traversal we now only need to find the matching sibling or insert a new child.

Creating the failure message is simply the travel from the current node to the root node in reverse order.