Reduce usage of IsSameOrEqualTo
`IsSameOrEqualTo` is necessary when comparing boxed numerics, e.g. that `1` is equal to `1L`.
`Type` and `XDocument` are not numerics, so this check is pure overhead and can be replaced with either `==` or `Equals`

`KeyValuePair.Equals` uses `ValueType.Equals(object)` and causes boxing.
This is avoided by converting the `KeyValuePair` into a 2-tuple and compare it against the default 2-tuple.