Make types generic in TAssertions
This PR makes most assertions generic in `TAssertions`.
This benefits deriving from one those assertion classes, as chaining
with `.And` allows you to continue use assertions from the derived class instead
of only from the base class.

Utilizing the generic type system this way also caught wrong return types.
`NullableDateTimeAssertions.Be` returned an
`DateTimeAssertions`, where it could return an
`NullableDateTimeAssertions`.
Similar "bugs" were caught in `NullableDateTimeOffsetAssertions` and
`NullableBooleanAssertions`.

Child classes are included to match the current signatures, which is
seen by noticing that all return types of `Should`s remain the same.
The only types that do not have child classes are
* `AndWhichConstraint`
* `DateTimeRangeAssertions`
* `DateTimeOffsetRangeAssertions`

This PR also served as a preparation for upcoming PRs that will improve
delegate and dictionary assertions