Improve test framework detection messages
If a user request an unsupported test framework via app settings, FA will now throw an exception instead of trying to detect a test framework with `AttemptToDetectUsingDynamicScanning`.

If the requested test framework is not available the failure message will contain the missing assembly name.

In both cases the failure message will contain the names of all supported frameworks.