Fix uncompilable async example
Since the introduction of `Awaiting(ValueTask)` `Awaiting(async => await
xxx)` has become ambiguous between the `Task` and `ValueTask` overloads.
This is not a bug in Fluent Assertions but a limitation in the compiler.
The best we can do for now is to update the documentation to avoid
confusing the compiler.