Hide protected member to avoid confusions for folks  upgrading to v6 (#1642)
In v6, there's no Should().Throw() that handles async code because you need to use Should().ThrowAsync() for that. However, while dogfooding, I discovered that existing calls to ThrowAsync() will raise a "Cannot access a protected member here". By renaming, it'll be clearer that a different method must be used.