Trailing slash results in 404 page
Noticed that when going to https://fluentassertions.com/introduction/ I get a 404-page in response. If I remove the trailing slash it works.
I guess the best solution is to make it work with and without trailing slash but this is at least one way of making the Issue template not pointing to a 404-page.