Simplify FrameworkSpecs (#2712)
While it makes sense not to take hard dependencies on test frameworks in the FluentAssertions library, the TestFrameworks specs all have a dependency on the their respective test frameworks. Use them to simplify the assertion on the exceptions which should be thrown.

Note that the support for NUnit4 was introduced exactly like this in 2f56211b769b7a9a10b619d8f3a2d868dd473735.